implement the ability to call us without an interface
As per the dbus specification, it is allowed to call a method without
specifying an interface. We do not allow it, however, and error with a console
log of "Interface undefined is not supported".

This commit adds an experimental implementation to allow method calls without
specifying an interface: We look through all interfaces, until a matching
method (by name) is found, and then call that method.

The implementation is experimental, as it contains lots of console.log() statements, which should be cleaned out.